AT8-195 - Backend - Implementação inicial do módulo de relatório de entradas com o relatório de comprovantes mensais

+++ database/migrations/2023_05_28_000010_create_plans_table.php
@@ -19,7 +19,7 @@ class CreatePlansTable extends Migration
             $table->integer('id', true);
             $table->string('name')->unique()->nullable(false);
             $table->string('description')->nullable(false);
-            $table->float('price')->nullable(false);
+            $table->decimal('price')->nullable(false);
             $table->boolean('activated')->nullable(false);
 
             $table->timestamps();
